+++ packages/sit-onyx/src/components/OnyxInput/OnyxInput.vue
@@ -149,8 +149,6 @@ const shouldShowCounter = computed(() => props.withCounter && props.maxlength);
   --selection-color: var(--onyx-color-base-primary-200);
 
   font-family: var(--onyx-font-family);
-  display: block;
-
   display: flex;
   flex-direction: column;
   gap: var(--onyx-spacing-5xs);
@@ -234,6 +232,7 @@ const shouldShowCounter = computed(() => props.withCounter && props.maxlength);
 
     &::placeholder {
       color: var(--onyx-color-text-icons-neutral-soft);
+      font-weight: 400;
       opacity: 1;
     }