+++ packages/chartjs-plugin/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @sit-onyx/chartjs-plugin
 
+## 1.0.0-alpha.13
+
+### Patch Changes
+
+- Updated dependencies [70b545e]
+  - sit-onyx@1.0.0-alpha.58
+
+## 1.0.0-alpha.12
+
+### Patch Changes
+
+- Updated dependencies [d4bb972]
+  - sit-onyx@1.0.0-alpha.57
+
+## 1.0.0-alpha.11
+
+### Patch Changes
+
+- 94e56ee: fix: update color value of `getDatasetColor()` when switching between light/dark mode
+- Updated dependencies [6631b3d]
+  - sit-onyx@1.0.0-alpha.56
+
 ## 1.0.0-alpha.10
 
 ### Patch Changes